My task 6 commit

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ long int b;
 long long int c;
 char d;
 float f;
-printf("size of a char: 1 byte(s)\n", (unsigned long)sizeof(d));
-printf("size of an int: 4 byte(s)\n", (unsigned long)sizeof(a));
-printf("size of a long int: 4 byte(s)\n", (unsigned long)sizeof(b));
-printf("size of a long long int: 8 byte(s)\n", (unsigned long)sizeof(c));
-printf("size of a float: 4 byte(s)\n", (unsigned long)sizeof(f));
+printf("size of a char: 1 byte(s)\n", sizeof(d));
+printf("size of an int: 4 byte(s)\n", sizeof(a));
+printf("size of a long int: 4 byte(s)\n", sizeof(b));
+printf("size of a long long int: 8 byte(s)\n", sizeof(c));
+printf("size of a float: 4 byte(s)\n", sizeof(f));
 return (0);
 }
